Split public config projection helpers
